frame/server/requirements.txt
attrs==21.4.0
click==8.1.2
Flask==2.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.4.0
loguru==0.6.0
MarkupSafe==2.1.1
peewee==3.14.10
PyJWT==2.3.0
pyrsistent==0.18.1
PyYAML==6.0
six==1.16.0
Werkzeug==2.1.1
